recon_requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.2.1
bitsandbytes==0.45.2
blinker==1.9.0
boto3==1.36.24
botocore==1.36.24
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==39.0.2
cycler==0.12.1
deepspeed==0.14.5
diffusers==0.30.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
flash-attn==2.7.1.post4
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
latex2mathml==3.77.0
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
megfile==4.1.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.27.1
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
orjson==3.10.15
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
peft==0.14.0
pillow==10.4.0
piper_sdk==0.2.19
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-can==4.5.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.6
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.2.2
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.22.0
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.1.2
torchdiffeq==0.2.5
torchvision==0.16.2
tqdm==4.67.1
transformers==4.43.1
triton==2.1.0
typer==0.15.1
typing_extensions==4.13.2
tzdata==2025.1
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.6
wavedrom==2.0.3.post3
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
